refactor: move motd modal
