+++ frontend/src/components/global-dialogs/motd-modal/motd-modal.tsx
@@ -6,10 +6,10 @@
 import { cypressId } from '../../../utils/cypress-attribute'
 import { Logger } from '../../../utils/logger'
 import { testId } from '../../../utils/test-id'
+import { CommonModal } from '../../common/modals/common-modal'
+import { RendererIframe } from '../../common/renderer-iframe/renderer-iframe'
 import { EditorToRendererCommunicatorContextProvider } from '../../editor-page/render-context/editor-to-renderer-communicator-context-provider'
 import { RendererType } from '../../render-page/window-post-message-communicator/rendering-message'
-import { CommonModal } from '../modals/common-modal'
-import { RendererIframe } from '../renderer-iframe/renderer-iframe'
 import { fetchMotd, MOTD_LOCAL_STORAGE_KEY } from './fetch-motd'
 import React, { useCallback, useEffect, useMemo, useState } from 'react'
 import { Button, Modal } from 'react-bootstrap'
